refactor(cli): simplify profile reporting structure
