+++ prof/jit.go
@@ -1,10 +1,5 @@
 package prof
 
-import (
-	"sort"
-	"strconv"
-)
-
 // Frontend identifies the compiler frontend that produced native code.
 type Frontend uint8
 
@@ -32,49 +27,6 @@ type ExitReason uint8
 // Counter is a pre-resolved runtime metric handle.
 type Counter struct {
 	value uint64
-}
-
-type jitMetrics struct {
-	captures map[captureKey]*Counter
-	compiles map[compileKey]*Counter
-	emits    map[entryKey]*emitCounters
-	entries  map[entryKey]*Counter
-	yields   map[entryKey]*Counter
-	exits    map[exitKey]*Counter
-}
-
-type captureKey struct {
-	fn      int
-	ip      int
-	outcome CaptureOutcome
-	reason  CaptureReason
-}
-
-type compileKey struct {
-	fn       int
-	ip       int
-	trigger  Trigger
-	frontend Frontend
-	outcome  CompileOutcome
-	reason   CompileReason
-}
-
-type entryKey struct {
-	fn       int
-	ip       int
-	kind     EntryKind
-	frontend Frontend
-}
-
-type exitKey struct {
-	entryKey
-	reason ExitReason
-	opcode int
-}
-
-type emitCounters struct {
-	emits Counter
-	bytes Counter
 }
 
 const (
@@ -223,336 +175,4 @@ func (c *Collector) RegisterYield(fn, ip int, kind EntryKind, frontend Frontend)
 // Inc increments a pre-resolved counter without allocating.
 func (c *Counter) Inc() {
 	c.value++
-}
-
-func (m *jitMetrics) appendMetrics(out []Metric, collector *Collector) []Metric {
-	captures := activeKeys(m.captures)
-	sort.Slice(captures, func(i, j int) bool { return captures[i].less(captures[j]) })
-	for _, key := range captures {
-		out = append(out, Metric{
-			Name: "vm_jit_trace_captures_total",
-			Labels: []Label{
-				{Key: "func", Value: strconv.Itoa(key.fn)},
-				{Key: "ip", Value: strconv.Itoa(key.ip)},
-				{Key: "outcome", Value: key.outcome.label()},
-				{Key: "reason", Value: key.reason.label()},
-			},
-			Value: float64(m.captures[key].value),
-		})
-	}
-
-	compiles := activeKeys(m.compiles)
-	sort.Slice(compiles, func(i, j int) bool { return compiles[i].less(compiles[j]) })
-	for _, key := range compiles {
-		out = append(out, Metric{
-			Name: "vm_jit_compiles_total",
-			Labels: []Label{
-				{Key: "func", Value: strconv.Itoa(key.fn)},
-				{Key: "ip", Value: strconv.Itoa(key.ip)},
-				{Key: "trigger", Value: key.trigger.label()},
-				{Key: "frontend", Value: key.frontend.label()},
-				{Key: "outcome", Value: key.outcome.label()},
-				{Key: "reason", Value: key.reason.label()},
-			},
-			Value: float64(m.compiles[key].value),
-		})
-	}
-
-	emits := activeEmits(m.emits)
-	sort.Slice(emits, func(i, j int) bool { return emits[i].less(emits[j]) })
-	for _, key := range emits {
-		labels := key.labels()
-		counters := m.emits[key]
-		if counters.emits.value > 0 {
-			out = append(out, Metric{Name: "vm_jit_entry_emits_total", Labels: labels, Value: float64(counters.emits.value)})
-		}
-		if counters.bytes.value > 0 {
-			out = append(out, Metric{Name: "vm_jit_entry_bytes_total", Labels: key.labels(), Value: float64(counters.bytes.value)})
-		}
-	}
-
-	out = appendEntryMetrics(out, "vm_jit_native_entries_total", m.entries)
-	out = appendEntryMetrics(out, "vm_jit_native_yields_total", m.yields)
-
-	exits := activeKeys(m.exits)
-	sort.Slice(exits, func(i, j int) bool { return exits[i].less(exits[j]) })
-	for _, key := range exits {
-		opcode := "none"
-		if key.opcode >= 0 && key.opcode <= 255 {
-			opcode = collector.opcodeLabel(byte(key.opcode))
-		}
-		out = append(out, Metric{
-			Name: "vm_jit_native_exits_total",
-			Labels: []Label{
-				{Key: "func", Value: strconv.Itoa(key.fn)},
-				{Key: "ip", Value: strconv.Itoa(key.ip)},
-				{Key: "kind", Value: key.kind.label()},
-				{Key: "frontend", Value: key.frontend.label()},
-				{Key: "reason", Value: key.reason.label()},
-				{Key: "opcode", Value: opcode},
-			},
-			Value: float64(m.exits[key].value),
-		})
-	}
-	return out
-}
-
-func (m *jitMetrics) merge(other *jitMetrics) {
-	mergeCounters(&m.captures, other.captures)
-	mergeCounters(&m.compiles, other.compiles)
-	if len(other.emits) > 0 && m.emits == nil {
-		m.emits = make(map[entryKey]*emitCounters)
-	}
-	for key, source := range other.emits {
-		if source.emits.value == 0 && source.bytes.value == 0 {
-			continue
-		}
-		destination := m.emits[key]
-		if destination == nil {
-			destination = &emitCounters{}
-			m.emits[key] = destination
-		}
-		destination.emits.value += source.emits.value
-		destination.bytes.value += source.bytes.value
-	}
-	mergeCounters(&m.entries, other.entries)
-	mergeCounters(&m.yields, other.yields)
-	mergeCounters(&m.exits, other.exits)
-}
-
-func (m *jitMetrics) reset() {
-	resetCounters(m.captures)
-	resetCounters(m.compiles)
-	for _, counters := range m.emits {
-		counters.emits.value = 0
-		counters.bytes.value = 0
-	}
-	resetCounters(m.entries)
-	resetCounters(m.yields)
-	resetCounters(m.exits)
-}
-
-func (k captureKey) less(other captureKey) bool {
-	return k.fn < other.fn ||
-		k.fn == other.fn && (k.ip < other.ip ||
-			k.ip == other.ip && (k.outcome < other.outcome ||
-				k.outcome == other.outcome && k.reason < other.reason))
-}
-
-func (k compileKey) less(other compileKey) bool {
-	return k.fn < other.fn ||
-		k.fn == other.fn && (k.ip < other.ip ||
-			k.ip == other.ip && (k.trigger < other.trigger ||
-				k.trigger == other.trigger && (k.frontend < other.frontend ||
-					k.frontend == other.frontend && (k.outcome < other.outcome ||
-						k.outcome == other.outcome && k.reason < other.reason))))
-}
-
-func (k entryKey) labels() []Label {
-	return []Label{
-		{Key: "func", Value: strconv.Itoa(k.fn)},
-		{Key: "ip", Value: strconv.Itoa(k.ip)},
-		{Key: "kind", Value: k.kind.label()},
-		{Key: "frontend", Value: k.frontend.label()},
-	}
-}
-
-func (k entryKey) less(other entryKey) bool {
-	return k.fn < other.fn ||
-		k.fn == other.fn && (k.ip < other.ip ||
-			k.ip == other.ip && (k.kind < other.kind ||
-				k.kind == other.kind && k.frontend < other.frontend))
-}
-
-func (k exitKey) less(other exitKey) bool {
-	return k.entryKey.less(other.entryKey) ||
-		k.entryKey == other.entryKey && (k.reason < other.reason ||
-			k.reason == other.reason && k.opcode < other.opcode)
-}
-
-func (v Frontend) label() string {
-	switch v {
-	case FrontendStatic:
-		return "static"
-	case FrontendTrace:
-		return "trace"
-	default:
-		return "none"
-	}
-}
-
-func (v Trigger) label() string {
-	switch v {
-	case TriggerHot:
-		return "hot"
-	case TriggerSideExit:
-		return "side-exit"
-	default:
-		return "none"
-	}
-}
-
-func (v EntryKind) label() string {
-	switch v {
-	case EntryStart:
-		return "start"
-	case EntryCall:
-		return "call"
-	case EntryLoop:
-		return "loop"
-	default:
-		return "none"
-	}
-}
-
-func (v CaptureOutcome) label() string {
-	switch v {
-	case CaptureOutcomePublished:
-		return "published"
-	case CaptureOutcomePartial:
-		return "partial"
-	case CaptureOutcomeRejected:
-		return "rejected"
-	default:
-		return "none"
-	}
-}
-
-func (v CaptureReason) label() string {
-	switch v {
-	case CaptureReasonAttemptLimit:
-		return "attempt-limit"
-	case CaptureReasonInvalidAnchor:
-		return "invalid-anchor"
-	case CaptureReasonHostCall:
-		return "host-call"
-	case CaptureReasonTailClosure:
-		return "tail-closure"
-	case CaptureReasonUnsupportedOp:
-		return "unsupported-op"
-	case CaptureReasonNestedTerminal:
-		return "nested-terminal"
-	case CaptureReasonStepTrap:
-		return "step-trap"
-	case CaptureReasonOpLimit:
-		return "op-limit"
-	default:
-		return "none"
-	}
-}
-
-func (v CompileOutcome) label() string {
-	switch v {
-	case CompileOutcomeEmitted:
-		return "emitted"
-	case CompileOutcomeEmpty:
-		return "empty"
-	case CompileOutcomeRejected:
-		return "rejected"
-	case CompileOutcomeError:
-		return "error"
-	default:
-		return "none"
-	}
-}
-
-func (v CompileReason) label() string {
-	switch v {
-	case CompileReasonNoInput:
-		return "no-input"
-	case CompileReasonNoPlan:
-		return "no-plan"
-	case CompileReasonInvalidPlan:
-		return "invalid-plan"
-	case CompileReasonLoweringRejected:
-		return "lowering-rejected"
-	case CompileReasonRegisterPressure:
-		return "register-pressure"
-	case CompileReasonBranchRange:
-		return "branch-range"
-	case CompileReasonBackendUnavailable:
-		return "backend-unavailable"
-	case CompileReasonError:
-		return "error"
-	default:
-		return "none"
-	}
-}
-
-func (v ExitReason) label() string {
-	switch v {
-	case ExitGuardKind:
-		return "guard-kind"
-	case ExitGuardShape:
-		return "guard-shape"
-	case ExitGuardBounds:
-		return "guard-bounds"
-	case ExitGuardValue:
-		return "guard-value"
-	case ExitColdBranch:
-		return "cold-branch"
-	case ExitTraceCut:
-		return "trace-cut"
-	case ExitTerminalOp:
-		return "terminal-op"
-	case ExitLoop:
-		return "loop-exit"
-	default:
-		return "none"
-	}
-}
-
-func appendEntryMetrics(out []Metric, name string, rows map[entryKey]*Counter) []Metric {
-	keys := activeKeys(rows)
-	sort.Slice(keys, func(i, j int) bool { return keys[i].less(keys[j]) })
-	for _, key := range keys {
-		out = append(out, Metric{Name: name, Labels: key.labels(), Value: float64(rows[key].value)})
-	}
-	return out
-}
-
-func counterFor[K comparable](rows *map[K]*Counter, key K) *Counter {
-	if *rows == nil {
-		*rows = make(map[K]*Counter)
-	}
-	counter := (*rows)[key]
-	if counter == nil {
-		counter = &Counter{}
-		(*rows)[key] = counter
-	}
-	return counter
-}
-
-func activeKeys[K comparable](rows map[K]*Counter) []K {
-	keys := make([]K, 0, len(rows))
-	for key, counter := range rows {
-		if counter.value > 0 {
-			keys = append(keys, key)
-		}
-	}
-	return keys
-}
-
-func activeEmits(rows map[entryKey]*emitCounters) []entryKey {
-	keys := make([]entryKey, 0, len(rows))
-	for key, counters := range rows {
-		if counters.emits.value > 0 || counters.bytes.value > 0 {
-			keys = append(keys, key)
-		}
-	}
-	return keys
-}
-
-func mergeCounters[K comparable](destination *map[K]*Counter, source map[K]*Counter) {
-	for key, counter := range source {
-		if counter.value > 0 {
-			counterFor(destination, key).value += counter.value
-		}
-	}
-}
-
-func resetCounters[K comparable](rows map[K]*Counter) {
-	for _, counter := range rows {
-		counter.value = 0
-	}
 }
